requirements.txt
discord.py
pyyaml
sqlalchemy
boto3
Flask
python-dotenv